label: represent 'SplitQ_t.CoverSplitArea' as 'uint64_t's
This is a preparatory change to handle upcoming changes to `RectArea` that will
have it return a `uint64_t`.

Gitlab: #2371

+++ lib/label/split.q.h
@@ -36,7 +36,7 @@ extern "C" {
 typedef struct split_q_s {
     struct Branch BranchBuf[NODECARD + 1];
     struct Rect CoverSplit;
-    unsigned int CoverSplitArea;
+    uint64_t CoverSplitArea;
     struct PartitionVars Partitions[METHODS];
 } SplitQ_t;
